+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ### >>> CHANGELOG <<<
 
 ### Version 0.2.0:
+- Added TlgTexture.Lock/Unlock/GetPixel/SetPixel()
+- Added TlgWindow.GetPixel/SetPixel()
+- Added pocketlang
 - Added first draft docs (LGT.chm, LGT.pdf)
 - Add a Sleep(0) to the while loop in TlgTaskList.Start()
 - Update TlgTexture.DrawTiled(), replace  AViewportWidth, AViewportHeight with AWindow: TlgWindow and get the viewport from